Refactoring a whole bunch of stuff. I was bad at committing, sorry
