+++ lib/articles.js
@@ -76,6 +76,7 @@ const LIST_ARTICLES_BY_AUTHOR = `
         headline
         byline
         slug
+        content
         authorSlugs
         authors {
           id
@@ -163,7 +164,7 @@ const GET_AUTHOR_BY_SLUG = `query getAuthor($slug: String) {
       slug
       name
       title
-      twitter 
+      twitter
       bio
       staff
       photoUrl
@@ -348,6 +349,15 @@ export async function listAllArticlesByTag(tag) {
       }
     }
   });
+
+  articlesByTag.map((article) => {
+    try {
+      article.content = JSON.parse(article.content);
+    } catch (e) {
+      console.log('error parsing json:', e);
+    }
+  });
+
   return articlesByTag;
 }
 
@@ -367,6 +377,15 @@ export async function listAllArticlesByAuthor(authorSlug) {
     LIST_ARTICLES_BY_AUTHOR,
     variables
   );
+
+  articlesData.content.data.map((article) => {
+    try {
+      article.content = JSON.parse(article.content);
+    } catch (e) {
+      console.log('error parsing json:', e);
+    }
+  });
+
   return articlesData.content.data;
 }
