+++ lib/siteMetadata.js
@@ -1,8 +1,8 @@
 module.exports = {
   siteMetadata: {
-    shortName: `tinynewsco`,
+    shortName: `The Oaklyn Observer`,
     siteUrl: `https://tinynewsco.org/`,
-    homepageTitle: `the tiny news collective`,
+    homepageTitle: `The Oaklyn Observer`,
     homepageSubtitle: `a local news initiative`,
     subscribe: {
       title: `Subscribe`,
