+++ package.json
@@ -6,6 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
+    "purgecss": "purgecss --css styles/bulma.css --content pages/** components/** --output ./purge.css",
     "lint": "eslint ./components/** ./pages/** ./lib/** -c .eslintrc.json --fix --ext js,jsx",
     "lint:fix": "npm run lint -- --fix",
     "format": "prettier --write ./components/** ./pages/** ./lib/**"
@@ -24,10 +25,11 @@
   "dependencies": {
     "bulma": "^0.9.0",
     "date-fns": "^2.14.0",
-    "dateline": "^2.2.1",
     "formik": "^2.1.4",
     "graphql-request": "^2.0.0",
+    "journalize": "^2.4.0",
     "next": "9.5",
+    "purgecss": "^2.3.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-ga": "^3.1.2",
