+++ pages/authors/[slug].js
@@ -7,15 +7,15 @@ import {
   getAuthorBySlug,
 } from '../../lib/articles.js';
 import { siteMetadata } from '../../lib/siteMetadata.js';
-import ArticleNav from '../../components/nav/ArticleNav.js';
-import ArticleFooter from '../../components/nav/ArticleFooter.js';
+import GlobalNav from '../../components/nav/GlobalNav.js';
+import GlobalFooter from '../../components/nav/GlobalFooter.js';
 import { useAmp } from 'next/amp';
 
 export default function AuthorPage(props) {
   const isAmp = useAmp();
   return (
     <Layout meta={siteMetadata}>
-      <ArticleNav metadata={siteMetadata} sections={props.sections} />
+      <GlobalNav sections={props.sections} />
       <section className="section">
         <h1 className="title">Articles by {props.author.name}</h1>
         <div className="columns">
@@ -26,7 +26,7 @@ export default function AuthorPage(props) {
           </div>
         </div>
       </section>
-      <ArticleFooter metadata={siteMetadata} />
+      <GlobalFooter />
     </Layout>
   );
 }
